Create First_Iteration
For my first attempt I wanted to create a list of strings which contained related locations, in this regard I created a list of areas in and around cape town and wanted to ask the user to name one place in cape town and check if the place they named was in this area. I wanted to get the users input, check the answer and display the results.